+++ src/main/java/techcourse/myblog/domain/article/Article.java
@@ -1,12 +1,12 @@
-package techcourse.myblog.article;
+package techcourse.myblog.domain.article;
 
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import techcourse.myblog.comment.Comment;
-import techcourse.myblog.date.BaseEntity;
+import techcourse.myblog.domain.comment.Comment;
+import techcourse.myblog.domain.date.BaseEntity;
 import techcourse.myblog.exception.InvalidAuthorException;
-import techcourse.myblog.user.User;
+import techcourse.myblog.domain.user.User;
 
 import javax.persistence.*;
 import java.util.ArrayList;
