+++ src/main/java/techcourse/myblog/domain/date/BaseEntity.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.date;
+package techcourse.myblog.domain.date;
 
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
