+++ src/main/java/techcourse/myblog/service/LoginService.java
@@ -2,9 +2,9 @@ package techcourse.myblog.service;
 
 import org.springframework.stereotype.Service;
 import techcourse.myblog.exception.LoginException;
-import techcourse.myblog.service.dto.AuthenticationDto;
-import techcourse.myblog.user.User;
-import techcourse.myblog.user.UserRepository;
+import techcourse.myblog.domain.dto.AuthenticationDto;
+import techcourse.myblog.domain.user.User;
+import techcourse.myblog.domain.user.UserRepository;
 
 import java.util.Optional;
