+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -6,11 +6,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.article.Article;
+import techcourse.myblog.domain.article.Article;
 import techcourse.myblog.service.ArticleService;
 import techcourse.myblog.service.CommentService;
-import techcourse.myblog.service.dto.ArticleDto;
-import techcourse.myblog.user.User;
+import techcourse.myblog.domain.dto.ArticleDto;
+import techcourse.myblog.domain.user.User;
 
 import javax.servlet.http.HttpSession;
