+++ src/main/java/techcourse/myblog/web/UserController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.exception.NotValidUpdateUserInfoException;
 import techcourse.myblog.exception.NotValidUserInfoException;
 import techcourse.myblog.service.UserService;
-import techcourse.myblog.service.dto.UserDto;
-import techcourse.myblog.service.dto.UserUpdateRequestDto;
-import techcourse.myblog.user.User;
+import techcourse.myblog.domain.dto.UserDto;
+import techcourse.myblog.domain.dto.UserUpdateRequestDto;
+import techcourse.myblog.domain.user.User;
 
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
