+++ src/test/java/techcourse/myblog/domain/user/UserTest.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.user;
+package techcourse.myblog.domain.user;
 
 import org.junit.jupiter.api.Test;
