+++ src/test/java/techcourse/myblog/service/ArticleServiceTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import techcourse.myblog.article.ArticleRepository;
+import techcourse.myblog.domain.article.ArticleRepository;
 import techcourse.myblog.exception.NotFoundObjectException;
-import techcourse.myblog.service.dto.ArticleDto;
+import techcourse.myblog.domain.dto.ArticleDto;
 
 import javax.servlet.http.HttpSession;
 import java.util.Optional;
@@ -14,8 +14,8 @@ import java.util.Optional;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.BDDMockito.given;
-import static techcourse.myblog.article.ArticleTest.article;
-import static techcourse.myblog.user.UserTest.user;
+import static techcourse.myblog.domain.article.ArticleTest.article;
+import static techcourse.myblog.domain.user.UserTest.user;
 
 
 @SpringBootTest
